Moved Logo file to public folder
