+++ frontend/public/YapYapLogo_notext.png
