+++ frontend/src/components/Navbar.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { Link } from "react-router-dom"
-import duckLogo from '../assets/YapYapLogo_notext.png'
+import duckLogo from '../../public/YapYapLogo_notext.png'
 
 const Navbar = () => {
     return(
